Moved procedures from runtime/functions.c to src/procedures.c as they are more suited to write there. Fix indentation and variable declaration. Modified tests related to create user as creating user isn't supported in all versions. Add test file names in latest/schedule.
