+++ test/JDBC/expected/Test-sp_droprolemember-vu-verify.out
@@ -20,23 +20,10 @@ int
 ~~END~~
 
 
--- Test whether sp_droprolemember_u1 is rolemember of sp_droprolemember_r2
-SELECT IS_ROLEMEMBER('sp_droprolemember_r2', 'sp_droprolemember_u1')
+-- Throw an error when member doesn't exist
+Exec sp_droprolemember 'sp_addrolemember_r1', 'sp_addrolemember_r3';
 GO
-~~START~~
-int
-1
-~~END~~
+~~ERROR (Code: 33557097)~~
 
-
-EXEC sp_droprolemember 'sp_droprolemember_r2', 'sp_droprolemember_u1';
-GO
-
--- Test whether sp_droprolemember_u1 is rolemember of sp_droprolemember_r2
-SELECT IS_ROLEMEMBER('sp_droprolemember_r2', 'sp_droprolemember_u1')
-GO
-~~START~~
-int
-0
-~~END~~
+~~ERROR (Message: role "master_sp_addrole_r3" does not exists)~~
