+++ test/JDBC/input/storedProcedures/Test-sp_addrolemember-vu-cleanup.sql
@@ -1,10 +1,4 @@
 -- tsql
-DROP USER sp_addrolemember_u1;
-GO
-
-DROP LOGIN sp_addrolemember_l1;
-GO
-
 DROP ROLE sp_addrolemember_r2;
 GO
