+++ test/JDBC/input/storedProcedures/Test-sp_addrolemember-vu-prepare.sql
@@ -4,9 +4,3 @@ GO
 
 CREATE ROLE sp_addrolemember_r2;
 GO
-
-CREATE LOGIN sp_addrolemember_l1 WITH PASSWORD = '123';
-GO
-
-CREATE USER sp_addrolemember_u1 FOR LOGIN sp_addrolemember_l1;
-GO
