+++ test/JDBC/input/storedProcedures/Test-sp_droprolemember-vu-cleanup.sql
@@ -1,10 +1,4 @@
 -- tsql
-DROP USER sp_droprolemember_u1;
-GO
-
-DROP LOGIN sp_droprolemember_l1;
-GO
-
 DROP ROLE sp_droprolemember_r2;
 GO
