+++ test/JDBC/input/storedProcedures/Test-sp_droprolemember-vu-verify.sql
@@ -10,13 +10,6 @@ GO
 SELECT IS_ROLEMEMBER('sp_droprolemember_r1', 'sp_droprolemember_r2')
 GO
 
--- Test whether sp_droprolemember_u1 is rolemember of sp_droprolemember_r2
-SELECT IS_ROLEMEMBER('sp_droprolemember_r2', 'sp_droprolemember_u1')
-GO
-
-EXEC sp_droprolemember 'sp_droprolemember_r2', 'sp_droprolemember_u1';
-GO
-
--- Test whether sp_droprolemember_u1 is rolemember of sp_droprolemember_r2
-SELECT IS_ROLEMEMBER('sp_droprolemember_r2', 'sp_droprolemember_u1')
+-- Throw an error when member doesn't exist
+Exec sp_droprolemember 'sp_addrolemember_r1', 'sp_addrolemember_r3';
 GO
