+++ test/JDBC/upgrade/14_5/schedule
@@ -123,10 +123,6 @@ insteadof_nested_trigger_with_dml
 nested_trigger_inside_proc
 nested_trigger_with_dml
 triggers_with_transaction
-Test-sp_addrole
-Test-sp_addrolemember
-Test-sp_droprole
-Test-sp_droprolemember
 Test-sp_helpdbfixedrole
 Test-sp_helpsrvrolemember
 Test-sp_helpuser
